move dataset merging to base class
