+++ tellurium/plotting/engine_plotly.py
@@ -1,7 +1,6 @@
 from .engine import PlottingEngine, PlottingFigure, PlottingLayout, filterWithSelections
-import plotly, itertools, numpy as np
+import plotly, numpy as np
 from plotly.graph_objs import Scatter, Scatter3d, Layout, Data
-from functools import reduce
 
 class PlotlyFigure(PlottingFigure):
     def __init__(self, title=None, layout=PlottingLayout(), logx=False, logy=False, save_to_pdf=False):
@@ -22,26 +21,6 @@ class PlotlyFigure(PlottingFigure):
                 autorange: True,
             }
         return Layout(**kwargs)
-
-    def getMergedTaggedDatasets(self):
-        for datasets_for_tag in self.tagged_data.values():
-            x = reduce(lambda u,v: np.concatenate((u,[np.nan],v)), (dataset['x'] for dataset in datasets_for_tag))
-            y = reduce(lambda u,v: np.concatenate((u,[np.nan],v)), (dataset['y'] for dataset in datasets_for_tag))
-            # merge all datasets
-            result_dataset = datasets_for_tag[0] if datasets_for_tag else None
-            for dataset in datasets_for_tag:
-                result_dataset.update(dataset)
-            # use the concatenated values for x and y
-            result_dataset['x'] = x
-            result_dataset['y'] = y
-            if result_dataset is not None:
-                yield result_dataset
-
-    def getDatasets(self):
-        """ Get an iterable of all datasets."""
-        return itertools.chain(
-            self.getMergedTaggedDatasets(),
-            (dataset for dataset in self.xy_datasets if not 'tag' in dataset))
 
     def plot(self):
         """ Plot the figure. Call this last."""
